Completed TX_VALID test vectors
- Added support for the tx_valid.json and tx_invalid.json from the SV Node test suite.

+++ src/main/java/org/twostack/bitcoin4j/transaction/TransactionOutput.java
@@ -66,7 +66,7 @@ public class TransactionOutput {
         WriteUtils writer = new WriteUtils();
 
         //write satoshi value
-        writer.writeUint64LE(satoshis);
+        writer.writeInt64LE(satoshis.longValue());
 
         //write the locking script
         byte[] outputScript = _lockingScriptBuilder.getLockingScript().getProgram();
